renaming file to covalent-tests.ts
